fix(exercises): use a 64-bit contract for shipping-capacity and container-water
Both reference solutions did signed arithmetic in `int`:
- binary-search-answer-capacity accumulated the upper bound with `0` (int
  accumulator), overflowing before the search on large total weight;
- array-container-most-water computed `min(h[i],h[j]) * (j - i)` as int*int,
  overflowing before assignment.

Adopt the preferred 64-bit contract across every parity surface. `min_capacity`
and `most_water` now return `long long`; the capacity bounds/load and the water
area are 64-bit, with one operand promoted before the multiply
(`1LL * min(...) * (j - i)`) and a 64-bit accumulate seed (`0LL`). Invalid-input
behavior is now explicit and tested: empty weights or `days <= 0` -> 0; fewer
than two lines -> 0. Added boundary tests whose results exceed INT_MAX.

Updated in parity: starters, solutions, C++ tests, exercise.json required_tests,
exercise-catalog.ts entries, the Code Lab stdin wrappers, and both READMEs.

Verified with the sanitizer (ASan+UBSan) harness: both solutions pass, both
starters build and fail as expected. Catalog and Code Lab config parity tests,
typecheck, and lint all pass.

Part of #615

+++ exercises/array-container-most-water/starter/most_water.hpp
@@ -5,14 +5,16 @@
 // Rules:
 //  - `most_water(height)` returns the largest min(h[i], h[j]) * (j - i).
 //  - Two pointers from the ends; move the pointer at the shorter line inward.
-//  - Fewer than two lines hold no water (area 0).
+//  - Heights are non-negative; fewer than two lines hold no water (area 0).
+//  - The area can exceed INT_MAX, so return 64-bit (long long) and promote one
+//    operand before multiplying (e.g. `1LL * min(...) * (j - i)`).
 //
 // Only edit this file. Do not change the public interface or the tests.
 #pragma once
 
 #include <vector>
 
-inline int most_water(const std::vector<int>& height) {
+inline long long most_water(const std::vector<int>& height) {
   // TODO: two pointers; track the maximum area; move the shorter side inward.
   (void)height;
   return 0;
